Added a function to load the data in and add outside processing
Added a function to load the data in and add outside processing  - like get 3 vectora and a vaule out of a structure - to change based on the way the data is recorded. Also the plots of TCI MACD and DATA pops up seperate from the app to make analisys easier.